requirements.txt
fastapi[all]
pymongo
python-jose
cryptography
python-dotenv
passlib[bcrypt]